Extract patch classes into separate files
